Cleanup _CONF, LHASH, STACK API's.
Also move some of related CHANGES entries closer together.

For LHASH:
- Move stats into separate sub-structure.
- Don't shrink table when in a "doall" loop, so the undocumented
  [gs]et_down_load aren't necessary (they're now deprecated) and fix
  the documentation.
- Deprecate the "print stats" functions. Don't collect statistics when
  deprecated; this saves at least one InterlockedIncrement operation for
  every fetch.

For _CONF_xxx:
- The _CONF_xxx() API was never public (i.e., not exported in
  libcrypto.num).  Move the functions that are only used in one place
  to become static functions in that file. Functions that are used more
  than once have their declaration moved to conf_local.h; conf_api.h is
  now just #include of conf.h
- Inline _CONF_get_section
- Merge conf_api into conf_def

For LHASH and STACK:
- Use OPENSSL_STACK_ instead of sk_ and OPENSSL_LHASH_ instead of lh_
