+++ crypto/conf/build.info
@@ -1,4 +1,4 @@
 LIBS=../../libcrypto
 SOURCE[../../libcrypto]= \
-        conf_err.c conf_lib.c conf_api.c conf_def.c conf_mod.c \
+        conf_err.c conf_lib.c conf_def.c conf_mod.c \
         conf_mall.c conf_sap.c conf_ssl.c
