+++ test/lhash_test.c
@@ -189,10 +189,12 @@ static int test_stress(void)
     if (!TEST_int_eq(lh_int_num_items(h), n))
             goto end;
 
+#ifndef OPENSSL_NO_DEPRECATED_3_0
     TEST_info("hash full statistics:");
     OPENSSL_LH_stats_bio((OPENSSL_LHASH *)h, bio_err);
     TEST_note("hash full node usage:");
     OPENSSL_LH_node_usage_stats_bio((OPENSSL_LHASH *)h, bio_err);
+#endif
 
     /* delete in a different order */
     for (i = 0; i < n; i++) {
@@ -209,10 +211,12 @@ static int test_stress(void)
         OPENSSL_free(p);
     }
 
+#ifndef OPENSSL_NO_DEPRECATED_3_0
     TEST_info("hash empty statistics:");
     OPENSSL_LH_stats_bio((OPENSSL_LHASH *)h, bio_err);
     TEST_note("hash empty node usage:");
     OPENSSL_LH_node_usage_stats_bio((OPENSSL_LHASH *)h, bio_err);
+#endif
 
     testresult = 1;
 end:
