+++ util/missingmacro.txt
@@ -89,6 +89,7 @@ LHASH_COMP_FN(3)
 LHASH_DOALL_ARG_FN(3)
 LHASH_OF(3)
 DEFINE_LHASH_OF(3)
+LHASH_OF_TYPE_DEPRECATED(3)
 int_implement_lhash_doall(3)
 OBJ_create_and_add_object(3)
 OBJ_bsearch(3)
